Extract tab component
