+++ assets/src/components/propertiesPanel/header.tsx
@@ -21,6 +21,7 @@ import { deselectVehicle } from "../../state"
 import CloseButton from "../closeButton"
 import { RouteVariantName } from "../routeVariantName"
 import VehicleIcon, { Orientation, Size } from "../vehicleIcon"
+import Tab from "./tab"
 import { TabMode } from "./tabPanels"
 
 interface Props {
@@ -29,23 +30,6 @@ interface Props {
   tabMode: TabMode
   setTabMode: Dispatch<SetStateAction<TabMode>>
 }
-
-const TabStatusIcon = () => (
-  <svg className="m-tabs__tab-status-icon" width="14" height="14">
-    <circle
-      className="m-tabs__tab-status-icon-outer-circle"
-      cx="7"
-      cy="7"
-      r="6"
-    />
-    <circle
-      className="m-tabs__tab-status-icon-inner-circle"
-      cx="7"
-      cy="7"
-      r="4"
-    />
-  </svg>
-)
 
 const vehicleOrientation = (
   vehicle: VehicleOrGhost,
@@ -140,29 +124,6 @@ const directionName = (
 ): string => (route ? route.directionNames[directionId] : "")
 
 const nowInSeconds = (): number => Math.floor(Date.now() / 1000)
-
-const Tab = ({
-  tabName,
-  activeTab,
-  setActiveTab,
-}: {
-  tabName: TabMode
-  activeTab: TabMode
-  setActiveTab: Dispatch<SetStateAction<TabMode>>
-}) => {
-  const classes =
-    tabName === activeTab ? "m-tabs__tab m-tabs__tab--selected" : "m-tabs__tab"
-
-  const tabTitle = tabName[0].toUpperCase() + tabName.slice(1)
-  const clickCallback = () => setActiveTab(tabName)
-
-  return (
-    <li className={classes} onClick={clickCallback}>
-      <TabStatusIcon />
-      {tabTitle}
-    </li>
-  )
-}
 
 const Header = ({ vehicle, route, tabMode, setTabMode }: Props) => {
   const [{ ladderDirections, settings }, dispatch] = useContext(
